Refactor based on PR comments
