+++ lib/ts/recipe/dashboard/api/userdetails/userPasswordPut.ts
@@ -2,6 +2,8 @@ import { APIInterface, APIOptions } from "../../types";
 import STError from "../../../../error";
 import EmailPasswordRecipe from "../../../emailpassword/recipe";
 import EmailPassword from "../../../emailpassword";
+import ThirdPartyEmailPasswordRecipe from "../../../thirdpartyemailpassword/recipe";
+import ThirdPartyEmailPassword from "../../../thirdpartyemailpassword";
 import { FORM_FIELD_PASSWORD_ID } from "../../../emailpassword/constants";
 
 type Response =
@@ -32,13 +34,63 @@ export const userPasswordPut = async (_: APIInterface, options: APIOptions): Pro
         });
     }
 
-    let passwordFormFields = EmailPasswordRecipe.getInstanceOrThrowError().config.signUpFeature.formFields.filter(
-        (field) => field.id === FORM_FIELD_PASSWORD_ID
-    );
+    let recipeToUse: "emailpassword" | "thirdpartyemailpassword" | undefined;
 
-    if (passwordFormFields.length === 0) {
+    try {
+        EmailPasswordRecipe.getInstanceOrThrowError();
+        recipeToUse = "emailpassword";
+    } catch (_) {}
+
+    if (recipeToUse === undefined) {
+        try {
+            ThirdPartyEmailPasswordRecipe.getInstanceOrThrowError();
+            recipeToUse = "thirdpartyemailpassword";
+        } catch (_) {}
+    }
+
+    if (recipeToUse === undefined) {
+        // This means that neither emailpassword or thirdpartyemailpassword is initialised
         throw new Error("Should never come here");
     }
+
+    if (recipeToUse === "emailpassword") {
+        let passwordFormFields = EmailPasswordRecipe.getInstanceOrThrowError().config.signUpFeature.formFields.filter(
+            (field) => field.id === FORM_FIELD_PASSWORD_ID
+        );
+
+        let passwordValidationError = await passwordFormFields[0].validate(newPassword);
+
+        if (passwordValidationError !== undefined) {
+            return {
+                status: "INVALID_PASSWORD_ERROR",
+                error: passwordValidationError,
+            };
+        }
+
+        const passwordResetToken = await EmailPassword.createResetPasswordToken(userId);
+
+        if (passwordResetToken.status === "UNKNOWN_USER_ID_ERROR") {
+            // Techincally it can but its an edge case so we assume that it wont
+            throw new Error("Should never come here");
+        }
+
+        const passwordResetResponse = await EmailPassword.resetPasswordUsingToken(
+            passwordResetToken.token,
+            newPassword
+        );
+
+        if (passwordResetResponse.status === "RESET_PASSWORD_INVALID_TOKEN_ERROR") {
+            throw new Error("Should never come here");
+        }
+
+        return {
+            status: "OK",
+        };
+    }
+
+    let passwordFormFields = ThirdPartyEmailPasswordRecipe.getInstanceOrThrowError().config.signUpFeature.formFields.filter(
+        (field) => field.id === FORM_FIELD_PASSWORD_ID
+    );
 
     let passwordValidationError = await passwordFormFields[0].validate(newPassword);
 
@@ -49,14 +101,17 @@ export const userPasswordPut = async (_: APIInterface, options: APIOptions): Pro
         };
     }
 
-    const passwordResetToken = await EmailPassword.createResetPasswordToken(userId);
+    const passwordResetToken = await ThirdPartyEmailPassword.createResetPasswordToken(userId);
 
     if (passwordResetToken.status === "UNKNOWN_USER_ID_ERROR") {
         // Techincally it can but its an edge case so we assume that it wont
         throw new Error("Should never come here");
     }
 
-    const passwordResetResponse = await EmailPassword.resetPasswordUsingToken(passwordResetToken.token, newPassword);
+    const passwordResetResponse = await ThirdPartyEmailPassword.resetPasswordUsingToken(
+        passwordResetToken.token,
+        newPassword
+    );
 
     if (passwordResetResponse.status === "RESET_PASSWORD_INVALID_TOKEN_ERROR") {
         throw new Error("Should never come here");
