+++ lib/ts/recipe/dashboard/utils.ts
@@ -28,7 +28,24 @@ import {
     USER_SESSIONS_API,
     VALIDATE_KEY_API,
 } from "./constants";
-import { APIInterface, RecipeIdForUser, RecipeInterface, TypeInput, TypeNormalisedInput } from "./types";
+import {
+    APIInterface,
+    EmailPasswordUser,
+    PasswordlessUser,
+    RecipeIdForUser,
+    RecipeInterface,
+    ThirdPartyUser,
+    TypeInput,
+    TypeNormalisedInput,
+} from "./types";
+import EmailPasswordRecipe from "../emailpassword/recipe";
+import ThirdPartyRecipe from "../thirdparty/recipe";
+import PasswordlessRecipe from "../passwordless/recipe";
+import EmailPassword from "../emailpassword";
+import ThirdParty from "../thirdparty";
+import Passwordless from "../passwordless";
+import ThirdPartyEmailPassword from "../thirdpartyemailpassword";
+import ThirdPartyPasswordless from "../thirdpartypasswordless";
 
 export function validateAndNormaliseUserInput(config: TypeInput): TypeNormalisedInput {
     if (config.apiKey.trim().length === 0) {
@@ -114,4 +131,149 @@ export function sendUnauthorisedAccess(res: BaseResponse) {
 
 export function isValidRecipeId(recipeId: string): recipeId is RecipeIdForUser {
     return recipeId === "emailpassword" || recipeId === "thirdparty" || recipeId === "passwordless";
+}
+
+export async function getUserForRecipeId(
+    userId: string,
+    recipeId: string
+): Promise<{
+    user: EmailPasswordUser | ThirdPartyUser | PasswordlessUser | undefined;
+    recipe:
+        | "emailpassword"
+        | "thirdparty"
+        | "passwordless"
+        | "thirdpartyemailpassword"
+        | "thirdpartypasswordless"
+        | undefined;
+}> {
+    let user: EmailPasswordUser | ThirdPartyUser | PasswordlessUser | undefined;
+    let recipe:
+        | "emailpassword"
+        | "thirdparty"
+        | "passwordless"
+        | "thirdpartyemailpassword"
+        | "thirdpartypasswordless"
+        | undefined;
+
+    if (recipeId === EmailPasswordRecipe.RECIPE_ID) {
+        try {
+            const userResponse = await EmailPassword.getUserById(userId);
+
+            if (userResponse !== undefined) {
+                user = {
+                    ...userResponse,
+                    firstName: "",
+                    lastName: "",
+                };
+                recipe = "emailpassword";
+            }
+        } catch (e) {
+            // No - op
+        }
+
+        if (user === undefined) {
+            try {
+                const userResponse = await ThirdPartyEmailPassword.getUserById(userId);
+
+                if (userResponse !== undefined) {
+                    user = {
+                        ...userResponse,
+                        firstName: "",
+                        lastName: "",
+                    };
+                    recipe = "thirdpartyemailpassword";
+                }
+            } catch (e) {
+                // No - op
+            }
+        }
+    } else if (recipeId === ThirdPartyRecipe.RECIPE_ID) {
+        try {
+            const userResponse = await ThirdParty.getUserById(userId);
+
+            if (userResponse !== undefined) {
+                user = {
+                    ...userResponse,
+                    firstName: "",
+                    lastName: "",
+                };
+                recipe = "thirdparty";
+            }
+        } catch (e) {
+            // No - op
+        }
+
+        if (user === undefined) {
+            try {
+                const userResponse = await ThirdPartyEmailPassword.getUserById(userId);
+
+                if (userResponse !== undefined) {
+                    user = {
+                        ...userResponse,
+                        firstName: "",
+                        lastName: "",
+                    };
+                    recipe = "thirdpartyemailpassword";
+                }
+            } catch (e) {
+                // No - op
+            }
+        }
+
+        if (user === undefined) {
+            try {
+                const userResponse = await ThirdPartyPasswordless.getUserById(userId);
+
+                if (userResponse !== undefined) {
+                    user = {
+                        ...userResponse,
+                        firstName: "",
+                        lastName: "",
+                    };
+                    recipe = "thirdpartypasswordless";
+                }
+            } catch (e) {
+                // No - op
+            }
+        }
+    } else if (recipeId === PasswordlessRecipe.RECIPE_ID) {
+        try {
+            const userResponse = await Passwordless.getUserById({
+                userId,
+            });
+
+            if (userResponse !== undefined) {
+                user = {
+                    ...userResponse,
+                    firstName: "",
+                    lastName: "",
+                };
+                recipe = "passwordless";
+            }
+        } catch (e) {
+            // No - op
+        }
+
+        if (user === undefined) {
+            try {
+                const userResponse = await ThirdPartyPasswordless.getUserById(userId);
+
+                if (userResponse !== undefined) {
+                    user = {
+                        ...userResponse,
+                        firstName: "",
+                        lastName: "",
+                    };
+                    recipe = "thirdpartypasswordless";
+                }
+            } catch (e) {
+                // No - op
+            }
+        }
+    }
+
+    return {
+        user,
+        recipe,
+    };
 }
